Character Scroller
This is a simple scroll routine that shifts text 1 character at a time. This is similiar to what one would achieve in Basic V2. It is not a smooth scroller.

Source is heavily documented and this source is for learning purposes.

I use screen frames to synchronise and slow down the scroll but you can also add a delay if you want which I have commented out.

Use the source freely.